enhance: improve performance and clean code
1. Add query has_contributor_location to avoid dup run
2. Check if location exists before analyze a repo to improve performance
3. skip auth
4. clean code for modulizing
