+++ github-handler/src/contributor_analysis.rs
@@ -1,10 +1,15 @@
 use chrono::{DateTime, FixedOffset};
-use model::github::ContributorAnalysis;
-use serde::{Deserialize, Serialize};
+use database::storage::Context;
+use model::github::{Contributor, ContributorAnalysis, GitHubUser};
 use std::collections::HashMap;
+use std::fs;
 use std::path::Path;
+use std::process::Command;
+use std::time::Duration;
 use tokio::process::Command as TokioCommand;
 use tracing::{debug, error, info, warn};
+
+use crate::{contributor_analysis, services::github_api::GitHubApiClient, BoxError};
 
 // 中国相关时区
 const CHINA_TIMEZONES: [&str; 4] = ["+0800", "+08:00", "CST", "Asia/Shanghai"];
@@ -44,6 +49,8 @@ pub async fn analyze_contributor_timezone(
     let mut timezone_count: HashMap<String, usize> = HashMap::new();
 
     // 分析每个提交的时区
+    // TODO: 是否有必要分析每个提交的时区，如果遇到一个就认为是中国，可能优化性能
+    // 如果遇到一个就认为是中国，可能优化性能
     for commit in &commits {
         let timezone = &commit.timezone;
 
@@ -78,7 +85,7 @@ struct CommitInfo {
     timezone: String,
 }
 
-/// 获取作者的所有提交
+/// 从git log里面获取作者的所有提交
 async fn get_author_commits(repo_path: &str, author_email: &str) -> Option<Vec<CommitInfo>> {
     let output = TokioCommand::new("git")
         .current_dir(repo_path)
@@ -126,126 +133,369 @@ async fn get_author_commits(repo_path: &str, author_email: &str) -> Option<Vec<C
     Some(commits)
 }
 
-/// 分析仓库的所有贡献者
-pub async fn analyze_repository_contributors(repo_path: &str) -> Vec<ContributorAnalysis> {
-    let mut results = Vec::new();
+// 分析贡献者国别位置
+async fn analyze_contributor_locations(
+    context: Context,
+    owner: &str,
+    repo: &str,
+    repository_id: &str,
+    contributors: &[Contributor],
+    github_users: &[GitHubUser],
+    email_to_user_id: &HashMap<String, i32>,
+) -> Result<(), BoxError> {
+    info!("分析仓库 {}/{} 的贡献者地理位置", owner, repo);
 
-    // 获取所有贡献者的邮箱
-    let emails = match get_all_contributor_emails(repo_path).await {
-        Some(emails) => emails,
-        None => {
-            error!("无法获取仓库贡献者邮箱: {}", repo_path);
-            return results;
-        }
-    };
-
-    info!("发现 {} 个贡献者邮箱", emails.len());
-
-    // 分析每个贡献者
-    for email in emails {
-        if let Some(analysis) = analyze_contributor_timezone(repo_path, &email).await {
-            debug!(
-                "分析完成: {} (可能来自中国: {})",
-                email,
-                if analysis.from_china { "是" } else { "否" }
-            );
-            results.push(analysis);
-        }
+    let base_dir = Path::new("/Users/Yetianxing/github_source");
+    if !base_dir.exists() {
+        fs::create_dir_all(base_dir)?;
+        info!("创建根目录: {:?}", base_dir);
     }
 
-    results
-}
+    // 构建目标路径: /mnt/crates/github_source/{owner}/{repo}
+    let target_dir = base_dir.join(format!("{}/{}", owner, repo));
+    let target_path = target_dir.to_string_lossy();
 
-/// 获取所有贡献者的邮箱
-async fn get_all_contributor_emails(repo_path: &str) -> Option<Vec<String>> {
-    let output = TokioCommand::new("git")
-        .current_dir(repo_path)
-        .args(["shortlog", "-sen", "HEAD"])
-        .output()
-        .await
-        .ok()?;
-
-    if !output.status.success() {
-        return None;
-    }
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let lines: Vec<&str> = stdout
-        .trim()
-        .split('\n')
-        .filter(|l| !l.is_empty())
-        .collect();
-
-    let mut emails = Vec::new();
-
-    for line in lines {
-        // 格式通常是: 123  Name <email@example.com>
-        if let Some(email_start) = line.find('<') {
-            if let Some(email_end) = line.find('>') {
-                let email = line[email_start + 1..email_end].trim().to_string();
-                emails.push(email);
+    // 检查目录是否已存在
+    if !target_dir.exists() {
+        // 确保父目录存在
+        if let Some(parent) = target_dir.parent() {
+            if !parent.exists() {
+                fs::create_dir_all(parent)?;
             }
         }
+
+        info!("克隆仓库到指定目录: {}", target_path);
+        let status = Command::new("git")
+            .args([
+                "clone",
+                "--config",
+                "credential.helper=reject", // 拒绝认证请求，不会提示输入
+                "--config",
+                "http.lowSpeedLimit=1000", // 设置低速限制
+                "--config",
+                "http.lowSpeedTime=10", // 如果速度低于限制持续10秒则失败
+                "--config",
+                "core.askpass=echo", // 不使用交互式密码提示
+                "--depth",
+                "1", // 浅克隆，只获取最近的提交
+                "--timeout",
+                "30", // 设置超时时间
+                &format!("https://github.com/{}/{}.git", owner, repo),
+                &target_path,
+            ])
+            .status();
+
+        match status {
+            Ok(status) if !status.success() => {
+                warn!("克隆仓库失败: {}，可能需要认证或不存在，跳过此仓库", status);
+                return Ok(());
+            }
+            Err(e) => {
+                warn!("执行git命令失败: {}，跳过此仓库", e);
+                return Ok(());
+            }
+            _ => {}
+        }
+    } else {
+        info!("更新已存在的仓库: {}", target_path);
+        let status = Command::new("git")
+            .current_dir(&target_dir)
+            .args([
+                "-c",
+                "credential.helper=reject",
+                "-c",
+                "http.lowSpeedLimit=1000",
+                "-c",
+                "http.lowSpeedTime=10",
+                "-c",
+                "core.askpass=echo",
+                "pull",
+                "--timeout",
+                "30",
+            ])
+            .status();
+
+        if let Err(e) = status {
+            warn!("更新仓库失败: {}，可能需要认证，继续分析当前代码", e);
+        }
     }
 
-    Some(emails)
-}
+    info!("开始分析 {} 个贡献者的时区信息", github_users.len());
 
-/// 生成仓库贡献者分析报告
-pub async fn generate_contributors_report(repo_path: &str) -> ContributorsReport {
-    info!("正在为仓库 {} 生成贡献者分析报告", repo_path);
-    let all_analyses = analyze_repository_contributors(repo_path).await;
+    let mut china_contributors = 0;
+    let mut non_china_contributors = 0;
 
-    // 获取中国贡献者和非中国贡献者的提交总数
-    let china_commits: usize = all_analyses.iter().filter(|c| c.from_china).count();
-    let non_china_commits: usize = all_analyses.len() - china_commits;
-    let total_commits = china_commits + non_china_commits;
+    // 对每个贡献者进行时区分析
+    for user in github_users.iter() {
+        // 使用贡献者的邮箱进行时区分析
+        let email = match &user.email {
+            Some(email) => email.clone(),
+            None => {
+                // 查找对应的contributor是否有邮箱
+                let contributor_email = contributors
+                    .iter()
+                    .find(|c| c.login == user.login)
+                    .and_then(|c| c.email.clone());
 
-    let china_percentage = if total_commits > 0 {
-        china_commits as f64 / total_commits as f64 * 100.0
+                match contributor_email {
+                    Some(email) => email,
+                    None => {
+                        warn!("用户 {} 没有邮箱信息，使用登录名作为替代", user.login);
+                        format!("{}@github.com", user.login)
+                    }
+                }
+            }
+        };
+
+        // 分析该贡献者的时区情况
+        let analysis =
+            match contributor_analysis::analyze_contributor_timezone(target_path.as_ref(), &email)
+                .await
+            {
+                Some(result) => result,
+                None => {
+                    warn!("无法分析用户 {} 的时区信息", user.login);
+                    continue;
+                }
+            };
+
+        // 查找用户ID
+        let user_id = match email_to_user_id.get(&email) {
+            Some(id) => *id,
+            None => match context
+                .github_handler_stg()
+                .get_user_id_by_name(&user.login)
+                .await
+            {
+                Ok(Some(id)) => id,
+                _ => {
+                    warn!("未找到用户 {} 的ID", user.login);
+                    continue;
+                }
+            },
+        };
+
+        // 存储贡献者位置分析
+        if let Err(e) = context
+            .github_handler_stg()
+            .store_contributor_location(repository_id, user_id, &analysis)
+            .await
+        {
+            error!("存储贡献者位置分析失败: {}", e);
+        }
+
+        // 统计中国贡献者和非中国贡献者
+        if analysis.from_china {
+            china_contributors += 1;
+            info!(
+                "贡献者 {} (邮箱: {}) 可能来自中国, 常用时区: {}",
+                user.login, email, analysis.common_timezone
+            );
+        } else {
+            non_china_contributors += 1;
+            info!(
+                "贡献者 {} (邮箱: {}) 可能来自海外, 常用时区: {}",
+                user.login, email, analysis.common_timezone
+            );
+        }
+    }
+
+    let total_contributors = china_contributors + non_china_contributors;
+    let china_percentage = if total_contributors > 0 {
+        (china_contributors as f64 / total_contributors as f64) * 100.0
     } else {
         0.0
     };
 
-    ContributorsReport {
-        total_contributors: all_analyses.len(),
-        china_contributors_count: china_commits,
-        non_china_contributors_count: non_china_commits,
+    info!(
+        "时区分析完成: 总计 {} 位贡献者, 其中中国贡献者 {} 位 ({:.1}%), 海外贡献者 {} 位 ({:.1}%)",
+        total_contributors,
+        china_contributors,
         china_percentage,
-        contributors: all_analyses,
+        non_china_contributors,
+        100.0 - china_percentage
+    );
+
+    // 查询中国贡献者统计
+    match context
+        .github_handler_stg()
+        .get_repository_china_contributor_stats(repository_id)
+        .await
+    {
+        Ok(stats) => {
+            info!(
+                "仓库 {}/{} 的中国贡献者统计: {}人中有{}人来自中国 ({:.1}%)",
+                owner,
+                repo,
+                stats.total_contributors,
+                stats.china_contributors,
+                stats.china_percentage
+            );
+
+            if !stats.china_contributors_details.is_empty() {
+                info!("中国贡献者TOP列表:");
+                for (i, contributor) in stats.china_contributors_details.iter().enumerate().take(5)
+                {
+                    let name_display = contributor
+                        .name
+                        .clone()
+                        .unwrap_or_else(|| contributor.login.clone());
+                    info!(
+                        "  {}. {} - {} 次提交",
+                        i + 1,
+                        name_display,
+                        contributor.contributions
+                    );
+                }
+            }
+        }
+        Err(e) => {
+            error!("获取中国贡献者统计失败: {}", e);
+        }
     }
+
+    Ok(())
 }
 
-/// Error type for contributor analysis
-type Result<T> = std::result::Result<T, Box<dyn std::error::Error + Send + Sync>>;
+// 分析Git贡献者
+pub(crate) async fn analyze_git_contributors(
+    context: Context,
+    owner: &str,
+    repo: &str,
+) -> Result<(), BoxError> {
+    info!("分析仓库贡献者: {}/{}", owner, repo);
 
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ContributorsReport {
-    pub total_contributors: usize,
-    pub china_contributors_count: usize,
-    pub non_china_contributors_count: usize,
-    pub china_percentage: f64,
-    pub contributors: Vec<ContributorAnalysis>,
-}
+    // 获取仓库ID
+    let repository_id = match context
+        .github_handler_stg()
+        .get_repository_id(owner, repo)
+        .await?
+    {
+        Some(id) => id,
+        None => {
+            warn!("仓库 {}/{} 未在数据库中注册", owner, repo);
+            return Ok(());
+        }
+    };
 
-impl ContributorsReport {
-    pub fn print_summary(&self) {
-        info!("贡献者分析报告摘要:");
-        info!("--------------------------------------------------");
-        info!("总贡献者: {} 人", self.total_contributors);
-        info!(
-            "中国贡献者: {} 人 ({:.1}%)",
-            self.china_contributors_count, self.china_percentage
-        );
-        info!(
-            "非中国贡献者: {} 人 ({:.1}%)",
-            self.non_china_contributors_count,
-            100.0 - self.china_percentage
-        );
-        info!("--------------------------------------------------");
+    // 查询是否存在贡献者位置信息
+    // 用于判断是否需要重新分析
+    match context
+        .github_handler_stg()
+        .has_contributor_location(&repository_id)
+        .await
+    {
+        Ok(true) => {
+            info!("仓库 {}/{} 已存在贡献者位置信息，跳过所有操作", owner, repo);
+            return Ok(());
+        }
+        Ok(false) => {
+            info!("仓库 {}/{} 没有贡献者位置信息，开始分析", owner, repo);
+        }
+        Err(e) => {
+            error!("查询仓库 {}/{} 的贡献者位置信息时出错: {}", owner, repo, e);
+            return Err(e.into());
+        }
     }
 
-    pub fn to_json(&self) -> Result<String> {
-        Ok(serde_json::to_string_pretty(self)?)
+    // 创建GitHub API客户端
+    let github_client = GitHubApiClient::new();
+
+    // 获取仓库贡献者
+    let contributors = github_client
+        .get_all_repository_contributors(owner, repo)
+        .await?;
+
+    info!("获取到 {} 个贡献者，开始存储到数据库", contributors.len());
+
+    // 使用HashMap存储邮箱到用户ID的映射，用于后续分析
+    let mut email_to_user_id = HashMap::new();
+    // 存储所有获取的用户信息，用于后续分析
+    let mut github_users = Vec::new();
+
+    // 存储贡献者信息
+    for contributor in &contributors {
+        // 获取并存储用户详细信息
+        let mut user = match github_client.get_user_details(&contributor.login).await {
+            Ok(user) => user,
+            Err(e) => {
+                warn!("获取用户 {} 详情失败: {}", contributor.login, e);
+                continue;
+            }
+        };
+
+        // 如果API返回的用户没有邮箱但贡献信息中有，则使用贡献中的邮箱
+        if user.email.is_none() && contributor.email.is_some() {
+            user.email = contributor.email.clone();
+        }
+
+        // 存储用户到数据库
+        let user_id = match context.github_handler_stg().store_user(&user).await {
+            Ok(id) => id,
+            Err(e) => {
+                error!("存储用户 {} 失败: {}", user.login, e);
+                continue;
+            }
+        };
+
+        // 保存邮箱到用户ID的映射
+        if let Some(email) = &user.email {
+            email_to_user_id.insert(email.clone(), user_id);
+            info!("记录邮箱映射: {} -> ID {}", email, user_id);
+        }
+
+        // 保存用户信息用于后续分析
+        github_users.push(user.clone());
+
+        // 存储贡献者关系
+        if let Err(e) = context
+            .github_handler_stg()
+            .store_contributor(&repository_id, user_id, contributor.contributions)
+            .await
+        {
+            error!(
+                "存储贡献者关系失败: {}/{} -> {}: {}",
+                owner, repo, user.login, e
+            );
+        }
+
+        // 等待一小段时间，避免触发GitHub API限制
+        tokio::time::sleep(Duration::from_millis(100)).await;
     }
+
+    // 查询并显示贡献者统计
+    match context
+        .github_handler_stg()
+        .query_top_contributors(&repository_id)
+        .await
+    {
+        Ok(top_contributors) => {
+            info!("仓库 {}/{} 的贡献者统计:", owner, repo);
+            for (i, contributor) in top_contributors.iter().enumerate().take(10) {
+                info!(
+                    "  {}. {} - {} 次提交",
+                    i + 1,
+                    contributor.login,
+                    contributor.contributions
+                );
+            }
+        }
+        Err(e) => {
+            error!("查询贡献者统计失败: {}", e);
+        }
+    }
+
+    // 分析贡献者国别 - 传递已获取的用户信息
+    analyze_contributor_locations(
+        context,
+        owner,
+        repo,
+        &repository_id,
+        &contributors,
+        &github_users,
+        &email_to_user_id,
+    )
+    .await?;
+
+    Ok(())
 }
